+++ ExampleMod/ExampleMod.cs
@@ -11,7 +11,8 @@ using ExampleMod.Common.Systems;
 
 namespace ExampleMod
 {
-	public class ExampleMod : Mod
+	// This is a partial class, meaning some of its parts were split into other files. See ExampleMod.*.cs for other portions.
+	public partial class ExampleMod : Mod
 	{
 		public const string AssetPath = "ExampleMod/Assets/";
 		public static ModKeybind RandomBuffKeybind;
@@ -26,56 +27,6 @@ namespace ExampleMod
 
 		public override void Unload() {
 			RandomBuffKeybind = null;
-		}
-
-		public override object Call(params object[] args) {
-			// The following code allows other mods to "call" Example Mod data.
-			// This allows mod developers to access Example Mod's data without having to set it a reference.
-			// Mod calls are not exposed by default, so it will be up to you to publish appropriate calls for your mod, and what values they return.
-
-			// Make sure the call doesn't include anything that could potentially cause exceptions.
-			if (args is null) {
-				throw new ArgumentNullException(nameof(args), "Arguments cannot be null!");
-			}
-
-			if (args.Length == 0) {
-				throw new ArgumentException("Arguments cannot be empty!");
-			}
-
-			// This check makes sure that the argument is a string using pattern matching.
-			// Since we only need one parameter, we'll take only the first item in the array.
-			if (args[0] is string content) {
-				// Returns the value provided by downedMinionBoss, if the argument calls for it.
-				if (content == "downedMinionBoss") {
-					return DownedBossSystem.downedMinionBoss;
-				}
-				// Returns the value provided by showMinionCount.
-				else if (content == "showMinionCount") {
-					return Main.LocalPlayer.GetModPlayer<ExampleInfoDisplayPlayer>().showMinionCount;
-				}
-
-				// Mod calls can also set data from our mod, as well as retrieve it.
-				if (content == "setMinionCount") {
-					// We need to make sure the call is provided with a value to set the field to.
-					if (args[1] is bool minionSet) {
-						Main.LocalPlayer.GetModPlayer<ExampleInfoDisplayPlayer>().showMinionCount = minionSet; // We'll set the value to what the argument provided.
-						// Optionally, you can return a value indicating that the assignment was successful.
-						// return true;
-					}
-					else { // If it's not the type we need, we can't continue
-						throw new Exception("Expected an argument of type bool when setting minion count, but got type " + args[1].GetType().Name + " instead."); // Tell the developer what type we need, and what we got instead.
-					}
-				}
-			}
-
-			// We can also do this with different data types.
-			if (args[0] is int contentInt && contentInt == 4) {
-				return ModContent.GetInstance<ExampleBiomeTileCount>().exampleBlockCount;
-			}
-
-			// If the arguments provided don't match anything we wanted to return a value for, we'll return null.
-			// This value can be anything you would like to provide as a default value.
-			return null;
 		}
 
 		//TODO: Introduce OOP packets into tML, to avoid this god-class level hardcode.
